Enum exercise.
Introduced Cats library with Dotty compat mode, to showcase good FP patterns.

Introduced a solutions project that depends on the main project. The idea is to put solution code into a namespaced object in solutions project, without having to duplicate all the code from the min one, as Ive learned from past workshops that leads to divergence with time.